+++ resources/views/bootstrap-components/form/file.blade.php
@@ -4,7 +4,9 @@
     @include('bootstrap-components::bootstrap-components.partials.label')
     @if($uploadedFileHtml->toHtml())
         {{ $uploadedFileHtml }}
-        @if($showRemoveCheckbox){{ bsCheckbox()->name('remove_' . $name )->containerClass(['mb-1']) }}@endif
+        @if($showRemoveCheckbox){{ bsCheckbox()->name('remove_' . $name )
+            ->label($removeCheckboxLabel)
+            ->containerClass(['mb-1']) }}@endif
     @endif
     <div class="input-group">
         @include('bootstrap-components::bootstrap-components.partials.icon')