+++ src/Form/Datetime.php
@@ -3,10 +3,11 @@
 namespace Okipa\LaravelBootstrapComponents\Form;
 
 use Carbon\Carbon;
-use Exception;
+use Okipa\LaravelBootstrapComponents\Form\Traits\DatetimeValidityChecks;
 
 class Datetime extends Input
 {
+    use DatetimeValidityChecks;
     /**
      * The component config key.
      *
@@ -56,34 +57,6 @@ class Datetime extends Input
         return array_merge(parent::values(), [
             'value' => $value,
         ]);
-    }
-
-    /**
-     * Check the component values validity
-     *
-     * @throws \Exception
-     * @return void
-     */
-    protected function checkValuesValidity(): void
-    {
-        parent::checkValuesValidity();
-        $this->value = $this->value ? $this->value : ($this->model ? $this->model->{$this->name} : null);
-        $this->format = $this->format ? $this->format : $this->defaultFormat();
-        if (! $this->format) {
-            throw new Exception(
-                get_class($this) . ' : No config or custom format is given for the bsDatetime() component.'
-            );
-        }
-        if ($this->value && is_string($this->value) && ! is_a($this->value, 'DateTime')) {
-            try {
-                Carbon::parse($this->value);
-            } catch (Exception $e) {
-                throw new Exception(
-                    get_class($this) . ' : the value must be a valid DateTime object or formatted string, « '
-                    . $this->value . ' » given.'
-                );
-            }
-        }
     }
 
     /**