+++ src/Form/Traits/DatetimeValidityChecks.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Okipa\LaravelBootstrapComponents\Form\Traits;
+
+use Carbon\Carbon;
+use Exception;
+
+trait DatetimeValidityChecks
+{
+    /**
+     * Check the component values validity
+     *
+     * @throws \Exception
+     * @return void
+     */
+    protected function checkValuesValidity(): void
+    {
+        parent::checkValuesValidity();
+        $this->value = $this->value ? $this->value : ($this->model ? $this->model->{$this->name} : null);
+        $this->format = $this->format ? $this->format : $this->defaultFormat();
+        if (! $this->format) {
+            throw new Exception(
+                get_class($this) . ' : No config or custom format is given for the bsDatetime() component.'
+            );
+        }
+        if ($this->value && is_string($this->value) && ! is_a($this->value, 'DateTime')) {
+            try {
+                Carbon::parse($this->value);
+            } catch (Exception $e) {
+                throw new Exception(
+                    get_class($this) . ' : the value must be a valid DateTime object or formatted string, « '
+                    . $this->value . ' » given.'
+                );
+            }
+        }
+    }
+}